% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimEpi_SIR.R
\name{SIRdataGen}
\alias{SIRdataGen}
\title{Stochastc SIR model (Osvaldo A.)}
\usage{
SIRdataGen(pop, nReps, nGens, repl, gen, r.g, r.f, TRbeta, RRgamma, SigG.g,
  SigG.f, SigE.g, SigE.f, sires, dpsire, N, gr.size, ngroups, freqDep, reprnum,
  SIRdata)
}
\arguments{
\item{pop}{the population data}

\item{nReps}{number of replicates}

\item{nGens}{number of generations}

\item{repl}{current replicate}

\item{gen}{current generation}

\item{r.g}{selection accuracy for susceptibility}

\item{r.f}{selection accuracy for infectivity}

\item{TRbeta}{average transmission coefficient beta}

\item{RRgamma}{recovery rate gamma}

\item{SigG.g}{genetic variance for susceptibility}

\item{SigG.f}{genetic variance for infectivity}

\item{SigE.g}{environmental variance for susceptibility}

\item{SigE.f}{environmental variance for infectivity}

\item{sires}{number of sires}

\item{dpsire}{number of dams per sire}

\item{N}{population size}

\item{gr.size}{group size}

\item{ngroups}{number of groups}

\item{freqDep}{T/F for frequency or density dependent force of infection}

\item{reprnum}{T/F modify calculation of infection rate to calculate reproduction number R0}

\item{SIRdata}{SIR data}
}
\value{
list with SIR data frame as time series (SIRts) and on a population level (pop)
}
\description{
Generating data from an stochastc SIR model
}
